Rendering the right page now
