+++ src/components/TakeInitiative/TakeInitiative.css
@@ -9,7 +9,7 @@ body {
 
 h1 {
   font-family: 'deutsch_gothicnormal', Arial, sans-serif;
-  font-size: 30px;
+  font-size: 45px;
   font-weight: bold;
   letter-spacing: 0.1rem;
   line-height: 0.5em;
@@ -19,7 +19,7 @@ h1 {
 
 @media (max-width: 400px) {
   h1 {
-    font-size: 1.5em;
+    font-size: 27px;
     letter-spacing: 0;
   }
 }
